Revert "Magic imports (#12)"
This reverts commit 32c53f3e14cd7a5a1422966435dee21e43031f0e.
